Structured the Samples into Chapters Signed-off-by: Dileep Miriyala <m.dileep@gmail.com>
