+++ src/NFugue.Samples/7.SaveAs/SaveAsStream.cs
@@ -7,19 +7,22 @@ using System.IO;
 
 namespace NFugue.Samples
 {
-	[Title("C#9:Save Pattern to Midi and Wave Stream")]
+	/// <summary>
+	/// Saves the pattern as Midi Memory Stream
+	/// </summary>
+	[Title(7, "Save Pattern to Midi and Wave Stream", 2)]
 	class SaveAsStream
 	{
 		public static void Run()
 		{
 			Pattern pattern = new ChordProgression("I IV V").Distribute("7%6")
-													   .AllChordsAs("$0 $0 $0 $0 $1 $1 $0 $0 $2 $1 $0 $0")
-													   .EachChordAs("$0ia100 $1ia80 $2ia80 $3ia80 $4ia100 $3ia80 $2ia80 $1ia80")
-													   .GetPattern()
-													   .SetInstrument(Instrument.AcousticBass)
-													   .SetTempo(100);
+															.AllChordsAs("$0 $0 $0 $0 $1 $1 $0 $0 $2 $1 $0 $0")
+															.EachChordAs("$0ia100 $1ia80 $2ia80 $3ia80 $4ia100 $3ia80 $2ia80 $1ia80")
+															.GetPattern()
+															.SetInstrument(Instrument.AcousticBass)
+															.SetTempo(100);
 			string midiFile = "TwelveBarBlues.mid";
-			
+
 			using (MemoryStream midiStream = new MemoryStream())
 			{
 				pattern.SaveAsMidi(midiStream);
