+++ src/NFugue.Samples/7.SaveAs/SaveAsWave.cs
@@ -7,22 +7,25 @@ using System.IO;
 
 namespace NFugue.Samples
 {
-	[Title("C#9:Save Pattern to Wave File")]
+	/// <summary>
+	/// Saves the pattern as a Wave file to the Disk
+	/// </summary>
+	[Title(7, "Save Pattern to Wave File", 3)]
 	class SaveAsWave
 	{
 		public static void Run()
 		{
 			Pattern pattern = new ChordProgression("I IV V").Distribute("7%6")
-													   .AllChordsAs("$0 $0 $0 $0 $1 $1 $0 $0 $2 $1 $0 $0")
-													   .EachChordAs("$0ia100 $1ia80 $2ia80 $3ia80 $4ia100 $3ia80 $2ia80 $1ia80")
-													   .GetPattern()
-													   .SetInstrument(Instrument.AcousticBass)
-													   .SetTempo(100);
+															.AllChordsAs("$0 $0 $0 $0 $1 $1 $0 $0 $2 $1 $0 $0")
+															.EachChordAs("$0ia100 $1ia80 $2ia80 $3ia80 $4ia100 $3ia80 $2ia80 $1ia80")
+															.GetPattern()
+															.SetInstrument(Instrument.AcousticBass)
+															.SetTempo(100);
 			string midiFile = "TwelveBarBlues.mid";
 
-			//pattern.SaveAsMidi(midiFile);
-			//MidiFileConverter.SavePatternToMidi(pattern, midiFile);
-			//WaveConverter.Convert(midiFile, midiFile + ".wav");
+			pattern.SaveAsMidi(midiFile);
+			MidiFileConverter.SavePatternToMidi(pattern, "F." + midiFile);
+			WaveConverter.Convert(midiFile, "F." + midiFile + ".wav");
 
 			using (MemoryStream midiStream = new MemoryStream())
 			{
