+++ src/NFugue.Samples/A.Parser/ParserDemo.cs
@@ -1,0 +1,33 @@
+﻿namespace NFugue.Samples
+{
+	/// <summary>
+	/// Connecting Any Parser to Any ParserListener
+	/// You can use JFugue to convert between music formats.
+	/// Most commonly, JFugue is used to turn Staccato music into MIDI sound.
+	/// Alternatively, you can play with the MIDI, MusicXML, and LilyPond parsers
+	/// and listeners.Or, you can easily create your own parser or listener,
+	/// and it will instantly interoperate with the other existing formats.
+	/// (And if you convert to Staccato, you can then play the Staccato music...and edit it!)
+	/// </summary>
+	/// <remarks>
+	/// Ported from https://github.com/dmkoelle/jfugue/blob/master/jfugue-sample 
+	///</remarks>
+	public class ParserDemo
+	{
+		public static void Run()
+		{
+			//String fileName = "PUT A MIDI FILE HERE";
+			//MidiParser parser = new MidiParser();
+			//StaccatoParserListener listener = new StaccatoParserListener();
+
+			//parser.AddParserListener(listener);
+			//parser.Parse(MidiSystem.getSequence(new File(fileName)));
+
+			//Pattern staccatoPattern = listener.getPattern();
+			//Console.WriteLine(staccatoPattern);
+
+			//Player player = new Player();
+			//player.Play(staccatoPattern);
+		}
+	}
+}
