+++ src/NFugue.Samples/C.Misc/MyParserListener.cs
@@ -1,12 +1,14 @@
 ﻿namespace NFugue.Samples
 {
-
-    class MyParserListener //: ParserListenerAdapter
+	/// <summary>
+	/// 
+	/// </summary>
+	/// <remarks>
+	/// Ported from https://github.com/dmkoelle/jfugue/blob/master/jfugue-sample 
+	///</remarks>
+	class MyParserListener //: ParserListenerAdapter
     {
-
         //public int counter;
-
-
         //public override void onNoteParsed(Note note)
         //{
         //    // A "C" note is in the 0th position of an octave
