--- src/NFugue.Samples/HelloWorld2.cs
@@ -1,26 +1,0 @@
-﻿using NFugue.Playing;
-using System;
-
-namespace NFugue.Samples
-{
-
-    /*	Playing multiple voices, multiple instruments, rests, chords, and durations
-    This example uses the Staccato 'V' command for specifing voices, 'I' for specifying instruments
-    (text within brackets is looked up in a dictionary and maps to MIDI instrument numbers),
-    '|' (pipe) for indicating measures (optional), durations including 'q' for quarter duration,
-    'qqq' for three quarter notes (multiple durations can be listed together), and 'h' for half,
-    'w' for whole, and '.' for a dotted duration; 'R' for rest, and the chords G-Major and C-Major.
-    Whitespace is not significant and can be used for visually pleasing or helpful spacing.
-    */
-    [Title("C#2:Hello World2")]
-    public class HelloWorld2
-    {
-        public static void Run()
-        {
-            using (Player player = new Player())
-            {
-                player.Play("V0 I[Flute] Eq Ch. | Eq Ch. | Dq Eq Dq Cq   V1 I[Flute] Rw | Rw | GmajQQQ CmajQ");
-            }
-        }
-    }
-}
