+++ src/NFugue.Samples/NFugue.Samples.csproj
@@ -75,33 +75,34 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdvancedChordProgressions.cs" />
-    <Compile Include="AdvancedRhythms.cs" />
-    <Compile Include="CarnaticReplacementMapDemo.cs" />
-    <Compile Include="SaveAsStream.cs" />
-    <Compile Include="SaveAsWave.cs" />
-    <Compile Include="SaveAsMidi.cs" />
-    <Compile Include="HelloWorld.cs" />
-    <Compile Include="HelloWorld2.cs" />
-    <Compile Include="IntroToChordProgressions.cs" />
-    <Compile Include="IntroToPatterns.cs" />
-    <Compile Include="IntroToPatterns2.cs" />
-    <Compile Include="IntroToRhythms.cs" />
-    <Compile Include="LSystemMusic.cs" />
-    <Compile Include="MyParserListener.cs" />
-    <Compile Include="ParserDemo.cs" />
-    <Compile Include="ParserDemo2.cs" />
+    <Compile Include="3.ChordProgressions\AdvancedChordProgressions.cs" />
+    <Compile Include="4.Rhythms\AdvancedRhythms.cs" />
+    <Compile Include="6.ReplacementMaps\CarnaticReplacementMapDemo.cs" />
+    <Compile Include="9.Try\PlayGround.cs" />
+    <Compile Include="7.SaveAs\SaveAsStream.cs" />
+    <Compile Include="7.SaveAs\SaveAsWave.cs" />
+    <Compile Include="7.SaveAs\SaveAsMidi.cs" />
+    <Compile Include="1.HelloWorld\HelloWorld.cs" />
+    <Compile Include="1.HelloWorld\HelloWorld2.cs" />
+    <Compile Include="3.ChordProgressions\IntroToChordProgressions.cs" />
+    <Compile Include="2.Patterns\IntroToPatterns.cs" />
+    <Compile Include="2.Patterns\IntroToPatterns2.cs" />
+    <Compile Include="4.Rhythms\IntroToRhythms.cs" />
+    <Compile Include="6.ReplacementMaps\LSystemMusic.cs" />
+    <Compile Include="C.Misc\MyParserListener.cs" />
+    <Compile Include="A.Parser\ParserDemo.cs" />
+    <Compile Include="A.Parser\ParserDemo2.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RealtimeExample.cs" />
-    <Compile Include="ReplacementMaps.cs" />
+    <Compile Include="B.RealTime\RealtimeExample.cs" />
+    <Compile Include="6.ReplacementMaps\ReplacementMaps.cs" />
     <Compile Include="SampleRunner.cs" />
-    <Compile Include="SeeMidi.cs" />
-    <Compile Include="SolfegeReplacementMapDemo.cs" />
-    <Compile Include="TemporalExample.cs" />
+    <Compile Include="8.Read\SeeMidi.cs" />
+    <Compile Include="6.ReplacementMaps\SolfegeReplacementMapDemo.cs" />
+    <Compile Include="C.Misc\TemporalExample.cs" />
     <Compile Include="TitleAttribute.cs" />
-    <Compile Include="TryThis.cs" />
-    <Compile Include="TwelveBarBlues.cs" />
+    <Compile Include="9.Try\TryThis.cs" />
+    <Compile Include="5.TweleveBar\TwelveBarBlues.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -115,6 +116,7 @@
       <Name>NFugue</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
