+++ src/NFugue.Samples/Program.cs
@@ -1,11 +1,10 @@
 ﻿namespace NFugue.Samples
 {
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            new SampleRunner().Run();
-
-        }
-    }
+	class Program
+	{
+		static void Main(string[] args)
+		{
+			new SampleRunner().Run();
+		}
+	}
 }
