+++ src/NFugue.Samples/TitleAttribute.cs
@@ -2,15 +2,18 @@
 
 namespace NFugue.Samples
 {
-    internal class TitleAttribute : Attribute
-    {
-        public TitleAttribute(string title, string description = "")
-        {
-            Title = title;
-            Description = description;
-        }
-
-        public string Title { get; set; }
-        public string Description { get; set; }
-    }
+	internal class TitleAttribute : Attribute
+	{
+		public TitleAttribute(int chapter, string title, int order = 0, string description = "")
+		{
+			Charpter = chapter;
+			Title = title;
+			Order = order;
+			Description = description;
+		}
+		public int Charpter { get; set; }
+		public int Order { get; set; }
+		public string Title { get; set; }
+		public string Description { get; set; }
+	}
 }
